src/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Bcrypt==1.0.1
python-decouple==3.8
Werkzeug==3.1.3
gunicorn==21.2.0
psycopg2-binary==2.9.9
openai==0.28.1
cryptography==41.0.7
